Created new utils folder for handling common tasks. Pulled firebase out of mlta-fonig and added to utils
