+++ modules/MLTA/mlta-config/package.json
@@ -14,7 +14,6 @@
   "dependencies": {
     "colors": "^1.1.2",
     "commander": "^2.9.0",
-    "firebase": "^3.4.1",
     "jsonfile": "^2.4.0",
     "mkdirp": "^0.5.1",
     "prompt": "^1.0.0",
