+++ modules/MLTA/utils/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "utils",
+  "version": "1.0.0",
+  "description": "Various utils for MLTA",
+  "main": "utils-firebase.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Alexander Clines",
+  "license": "ISC",
+  "dependencies": {
+    "firebase": "^3.4.1"
+  }
+}
